\name{cvq2.sample.C}
\alias{cvq2.sample.C}
\docType{data}
\title{
  Small data set to demonstrate the statistic methods.
}
\description{
  Contains a small data set with four observations and four predictions.
}
\usage{data(cvq2.sample.C)}
\format{
  A data frame with four observations and four predictions.
  \describe{
    \item{\code{observed}}{observation}
    \item{\code{predicted}}{prediction}
  }
}
%\details{}
\source{
  Generic data set, created for this purpose only.
}
\seealso{
  \code{\link[=cvq2-package]{cvq2}},
  \code{\link{predPow}}
}
%%\references{}
%\examples{
%data(cvq2.sample.C)
%}
\keyword{datasets}
